Rename asset page to capture details page.
