--- src/app/features/home/asset/asset.page.spec.ts
@@ -1,23 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { SharedTestingModule } from '../../../shared/shared-testing.module';
-import { AssetPage } from './asset.page';
-
-describe('AssetPage', () => {
-  let component: AssetPage;
-  let fixture: ComponentFixture<AssetPage>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [AssetPage],
-      imports: [SharedTestingModule],
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(AssetPage);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }));
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
