+++ src/app/features/home/capture-tab/capture-details/capture-details.module.ts
@@ -1,18 +1,18 @@
 import { NgModule } from '@angular/core';
-import { SharedModule } from '../../../shared/shared.module';
-import { AssetPageRoutingModule } from './asset-routing.module';
-import { AssetPage } from './asset.page';
+import { SharedModule } from '../../../../shared/shared.module';
+import { CaptureDetailsPageRoutingModule } from './capture-details-routing.module';
+import { CaptureDetailsPage } from './capture-details.page';
 import { ContactSelectionDialogComponent } from './contact-selection-dialog/contact-selection-dialog.component';
 import { FriendInvitationDialogComponent } from './contact-selection-dialog/friend-invitation-dialog/friend-invitation-dialog.component';
 import { OptionsMenuComponent } from './options-menu/options-menu.component';
 
 @NgModule({
-  imports: [SharedModule, AssetPageRoutingModule],
+  imports: [SharedModule, CaptureDetailsPageRoutingModule],
   declarations: [
-    AssetPage,
+    CaptureDetailsPage,
     ContactSelectionDialogComponent,
     FriendInvitationDialogComponent,
     OptionsMenuComponent,
   ],
 })
-export class AssetPageModule {}
+export class CaptureDetailsPageModule {}
