+++ src/app/features/home/capture-tab/capture-details/capture-details.page.html
