+++ src/app/features/home/capture-tab/capture-details/contact-selection-dialog/contact-selection-dialog.component.scss
