+++ src/app/features/home/capture-tab/capture-details/contact-selection-dialog/friend-invitation-dialog/friend-invitation-dialog.component.scss
