+++ src/app/features/home/capture-tab/capture-details/information/information-routing.module.ts
