+++ src/app/features/home/capture-tab/capture-details/information/information.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { SharedModule } from '../../../../shared/shared.module';
+import { SharedModule } from '../../../../../shared/shared.module';
 import { InformationPageRoutingModule } from './information-routing.module';
 import { InformationPage } from './information.page';
