+++ src/app/features/home/capture-tab/capture-details/options-menu/options-menu.component.ts
