+++ src/app/features/home/capture-tab/capture-details/sending-post-capture/sending-post-capture.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
-import { PostCaptureCardModule } from '../../../../shared/core/post-capture-card/post-capture-card.module';
-import { SharedModule } from '../../../../shared/shared.module';
+import { PostCaptureCardModule } from '../../../../../shared/core/post-capture-card/post-capture-card.module';
+import { SharedModule } from '../../../../../shared/shared.module';
 import { SendingPostCapturePageRoutingModule } from './sending-post-capture-routing.module';
 import { SendingPostCapturePage } from './sending-post-capture.page';
