+++ src/app/features/home/capture-tab/capture-details/sending-post-capture/sending-post-capture.page.html
