+++ src/app/features/home/capture-tab/capture-details/sending-post-capture/sending-post-capture.page.spec.ts
@@ -1,5 +1,5 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { SharedTestingModule } from '../../../../shared/shared-testing.module';
+import { SharedTestingModule } from '../../../../../shared/shared-testing.module';
 import { SendingPostCapturePage } from './sending-post-capture.page';
 
 describe('SendingPostCapturePage', () => {
