+++ src/app/features/home/capture-tab/capture-details/sending-post-capture/sending-post-capture.page.ts
@@ -11,19 +11,19 @@ import {
   shareReplay,
   switchMap,
 } from 'rxjs/operators';
-import { BlockingActionService } from '../../../../shared/services/blocking-action/blocking-action.service';
-import { ConfirmAlert } from '../../../../shared/services/confirm-alert/confirm-alert.service';
+import { BlockingActionService } from '../../../../../shared/services/blocking-action/blocking-action.service';
+import { ConfirmAlert } from '../../../../../shared/services/confirm-alert/confirm-alert.service';
 import {
   DiaBackendAsset,
   DiaBackendAssetRepository,
-} from '../../../../shared/services/dia-backend/asset/dia-backend-asset-repository.service';
-import { DiaBackendTransactionRepository } from '../../../../shared/services/dia-backend/transaction/dia-backend-transaction-repository.service';
-import { getOldProof } from '../../../../shared/services/repositories/proof/old-proof-adapter';
-import { ProofRepository } from '../../../../shared/services/repositories/proof/proof-repository.service';
+} from '../../../../../shared/services/dia-backend/asset/dia-backend-asset-repository.service';
+import { DiaBackendTransactionRepository } from '../../../../../shared/services/dia-backend/transaction/dia-backend-transaction-repository.service';
+import { getOldProof } from '../../../../../shared/services/repositories/proof/old-proof-adapter';
+import { ProofRepository } from '../../../../../shared/services/repositories/proof/proof-repository.service';
 import {
   isNonNullable,
   switchTap,
-} from '../../../../utils/rx-operators/rx-operators';
+} from '../../../../../utils/rx-operators/rx-operators';
 
 @UntilDestroy({ checkProperties: true })
 @Component({
